T4.0 support for FlexIO
At least some of it is working now.
I have not updated the background stuff yet that can update
on Flexio ISR code...

T4 does not have enough flexio pins on FlexIO1 nor FlexIO2 for this
And on FlexIO3 There is not 8 contiguous pins, but instead two nibbles with a two flxio pin gap.